refactor: change react project to preferred folder structure
