--- packages/client/src/App.tsx
@@ -1,29 +1,0 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
-
-import { product } from '@smdn-project/common'
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-      <p>{product.name}</p>
-    </div>
-  );
-}
-
-export default App;
