+++ packages/client/src/app/index.ts
@@ -1,0 +1,3 @@
+import App from './App'
+
+export default App
